[POC] Add support for passing commands to get username and password
This lets you use the CLI tool of your choice to supply username and
password, thereby letting you avoid storing your gmail username and
password in plain text in the config file.

POC because the code is brittle, ugly, undocumented, and untested.

That last is of course the major crime, and is because I got
authentication errors from Google Play no matter what I tried.

I tried two different versions of urllib as described in this issue:

https://github.com/matlink/gplaycli/issues/260

as well as using an app password rather than my usual normal password.

Alas, despite all that, I got nowhere, and so I threw my hands in the
air and gave up.